Fix pointer truncation to remove warning

+++ src/modinfodialog.cpp
@@ -1080,7 +1080,7 @@ void ModInfoDialog::openFile(const QModelIndex &index)
   QString fileName = m_FileSystemModel->filePath(index);
 
   HINSTANCE res = ::ShellExecuteW(nullptr, L"open", ToWString(fileName).c_str(), nullptr, nullptr, SW_SHOW);
-  if ((int)res <= 32) {
+  if ((unsigned long long)res <= 32) {
     qCritical("failed to invoke %s: %d", fileName.toUtf8().constData(), res);
   }
 }
